requirements.txt
Flask==3.0.2
pytest==8.0.2
google-generativeai
python-dotenv
